final changes

+++ delete_user.php
@@ -5,10 +5,17 @@ include_once('template.php');
 if (isset($_GET['id']) and isset($_SESSION['userId'])) {
 // Query that deletes using DELETE SQL statement.
 $query = <<<END
+SELCT * FROM project_Admin
+END;
+$res = $mysqli->query($query);
+if($res->num_rows > 1)
+{
+$query = <<<END
 DELETE FROM project_User
 WHERE UserID = '{$_GET['id']}'
 END;
 $mysqli->query($query);
+}
 header('Location:users.php');
 }
 echo $navigation;
